feat(tracing): add std/file modes and add controller.ProcessCSIMounts span for CSI remount after resume/upgrade
