+++ config/sandbox-manager/deployment.yaml
@@ -45,7 +45,7 @@ spec:
             # OpenTelemetry tracing (uncomment and configure to enable)
             # - --tracing-mode=otel
             # - --tracing-endpoint=otel-collector:4317
-            # - --tracing-insecure=true
+            # - --tracing-insecure=false
           ports:
             - containerPort: 8080
               name: http-manager
