+++ pkg/servers/e2b/snapshot.go
@@ -53,9 +53,13 @@ func (sc *Controller) CreateSnapshot(r *http.Request) (web.ApiResponse[*models.S
 			Message: fmt.Sprintf("Sandbox %s is not running", sandboxID),
 		}
 	}
+	// Capture the HTTP root span context before creating child spans, so that
+	// the trace context injected into the Checkpoint CR points at the root span
+	// and the controller-side spans attach to it correctly.
 	ctx = tracing.WithRootSpanContext(ctx)
 	ctx, span := tracing.Tracer("sandbox-manager").Start(ctx, tracing.SpanManagerCreateSnapshot)
 	defer span.End()
+	// Record optional request extensions as span attributes when present.
 	if request.Extensions.KeepRunning != nil {
 		span.SetAttributes(attribute.Bool(tracing.AttrSnapshotKeepRunning, *request.Extensions.KeepRunning))
 	}
